requirements.txt
# Core dependencies
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
PyYAML>=6.0
sentence-transformers>=2.2.0

# Web framework
fastapi>=0.100.0
uvicorn>=0.20.0

# UI framework
streamlit>=1.25.0

# Cloud dependencies (optional)
google-cloud-storage>=2.10.0
vertexai>=1.0.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0
